requirements.txt
Django==1.11
django-cors-headers==2.1.0
django-tastypie==0.14.0
djangorestframework==3.7.7
jdcal==1.0
Jinja2==2.8
pbr==1.8.1
pysqlite==2.8.3
python-dateutil==2.6.1
python-mimeparse==1.6.0
python-slugify==1.2.1
pytz==2015.7
PyYAML==3.12
pyzipcode==1.0
pyzmq==16.0.2
redis==2.10.3
requests==2.9.1
requests-toolbelt==0.6.2
rsa==3.4.2
s3transfer==0.1.12
scipy==0.18.1
selenium==3.0.1
semantic-version==2.5.0
singledispatch==3.4.0.3
six==1.11.0
slugify==0.0.1
SQLAlchemy==1.0.8
subprocess32==3.2.7
suds==0.4
unicodecsv==0.14.1
Unidecode==0.4.19
validate-email==1.3
vboxapi==1.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wcwidth==0.1.6
websocket-client==0.36.0
